refactor(api): split GitHub webhook delivery and add emulator coverage
Extract signature checks, allowlisted event metadata, and prompt framing
into github-webhook. Fence generic webhook text bodies as untrusted data.
Add a GitHub delivery emulator and route the API webhook suite through it
so CI proves HMAC minting and sanitized wakes without hand-rolled crypto.
